optimization

+++ src/model.py
@@ -97,7 +97,7 @@ class Model():
                     sys.exit()
                 self.embed_snt = self.last_src #[B,Hs*2]
             elif self.config.net_sentence == 'max':
-                self.mask1 = tf.sequence_mask(self.len_src)
+                self.mask1 = tf.sequence_mask(self.len_src, dtype=tf.float32)
                 self.mask2 = tf.cast(self.mask1, tf.float32) #tf.to_float(self.mask1) 
                 mask = tf.expand_dims(self.mask2, 2) #[B,Ss] => [B,Ss,1]
                 self.embed_snt = self.out_src * mask + (1-mask) * tf.float32.min #masked tokens contain -Inf
